address review: revert non-security-sensitive changes, fix modulo cast order
Revert crypto PRNG changes in syslog.go (fun messages), do.go (debug
command), and metricserver.go (best-effort shuffle) as these do not
need cryptographic randomness.

Fix stub_unsafe.go: apply modulo before int cast to avoid potential
negative index from uint64-to-int conversion.

+++ pkg/sentry/platform/systrap/stub_unsafe.go
@@ -239,7 +239,7 @@ func stubInit() {
 			if n == 0 {
 				panic("failed to map stub code")
 			}
-			r := regions[int(rand.Uint64())%n]
+			r := regions[rand.Uint64()%uint64(n)]
 			stubStart = r.start + uintptr(rand.Uint64())*hostarch.PageSize%(r.end-r.start-stubROMapEnd)
 		}
 		// Map the target address for the stub.
